change delete identity flow
